+++ test/functional_basics.js
@@ -1,7 +1,11 @@
 // Testing the basic functions of the functional testing API.
 
-const {test,plan,ok,pass,is,isnt,cmp,tap} = require('../lib').functional();
+const 
+{
+  test,plan,ok,pass,is,isnt,cmp,tap,isJSON,isntJSON,lives,
+} = require('../lib').functional();
 const def = require('./inc/basics.js');
+const stringify = def.types.stringify;
 
 plan(def.plan);
 
@@ -10,20 +14,37 @@ pass('pass()');
 
 for (const it of def.isTests)
 {
-  const t = JSON.stringify(it);
+  const t = stringify(it);
   is(it, it, 'is('+t+','+t+')');
 }
 
 for (const [a,b] of def.isntTests)
 {
-  const msg = `isnt(${JSON.stringify(a)},${JSON.stringify(b)})`;
+  const msg = `isnt(${stringify(a)},${stringify(b)})`;
   isnt(a, b, msg);
 }
 
 for (const args of def.cmpTests)
 {
-  const msg = 'cmp('+JSON.stringify(args)+')';
+  const msg = 'cmp('+stringify(args)+')';
   cmp(...args, msg);
+}
+
+for (const func of def.livesTests)
+{
+  lives(func, `lives(${stringify(func)})`);
+}
+
+for (const [a,b] of def.isJsonTests)
+{
+  const msg = `isJSON(${stringify(a)},${stringify(b)})`;
+  isJSON(a, b, msg);
+}
+
+for (const [a,b] of def.isntJsonTests)
+{
+  const msg = `isntJSON(${stringify(a)},${stringify(b)})`;
+  isntJSON(a, b, msg);
 }
 
 // We're done, output the log.